move gui + display multiple
