+++ src/main.py
@@ -1,20 +1,12 @@
-# from src.modules.data_processing import process_data
-# from src.gui.gui import start_gui
-
 import os
 import subprocess
 import torch
 from segment_anything import modeling, sam_model_registry, SamAutomaticMaskGenerator
-import cv2
-import matplotlib
-matplotlib.use("WebAgg")
-import tempfile
-import webbrowser
 
-import matplotlib.pyplot as plt, mpld3
-import matplotlib.axes as Axes
 from dotenv import load_dotenv
 from src.modules.sample import Sample
+from src.gui import display
+
 import sys
 
 
@@ -41,65 +33,23 @@ def init_mask_generator() -> SamAutomaticMaskGenerator:
     sam = init_sam()
     return SamAutomaticMaskGenerator(sam)
 
-def display_grid_of_all_polygons(ax: Axes, masks):
-    ax.axis('off')
-    ax.set_title("polygons")
-    # display grid of all masks
-    for i in range(len(masks)):
-        x, y = masks[i].polygon.exterior.xy
-        # rot = transforms.Affine2D().rotate_deg(90)
-        ax.plot(x, y, linewidth=1, color=(masks[i].color[0] / 255, masks[i].color[1] / 255, masks[i].color[2] / 255))
-        ax.yaxis.set_inverted(True)
-        ax.set_aspect('equal')
-
-
-def display_grid_of_all(ax: Axes, img, masks):
-    ax.set_title("segmentation")
-    ax.axis('off')
-    for mask in masks:
-        contour = max(mask.contours, key = cv2.contourArea)
-        annotated = cv2.drawContours(image=img, contours=[contour], contourIdx=0, color=mask.color, thickness=2)
-    ax.imshow(annotated)
-
-def corners_number_dist(ax, polygons):
-    corners = [len(p.exterior.coords) - 1 for p in polygons]
-    ax.hist(corners, bins=range(2, 12, 1), label="corners number")
-    ax.set_title("corners number dist")
-
-def area_dist(ax, polygons):
-    areas = [p.area for p in polygons]
-    ax.hist(areas, bins=range(0, int(max(areas)), 100), label="area dist")
-    ax.set_title("area dist")
-
-def average_angle_dist(ax, masks):
-    angles = [m.average_angle_per_polygon() for m in masks]
-    ax.hist(angles, bins=range(0, 180, 10), label="angle dist")
-    ax.set_title("largest angle diff dist")
 
 def main():
-    # setup()
-
     mask_generator = init_mask_generator()
-    sample = Sample(f"1.png", mask_generator)
-    polygons = [mask.polygon for mask in sample.masks]
-    print(f"setup done: {len(polygons)} polygons")
-    fig, axs = plt.subplot_mosaic([['A', 'A', 'B', 'B'],['A', 'A', 'B', 'B'], ['C', 'D', 'E', 'H']])
-    display_grid_of_all_polygons(axs['A'], sample.masks)
-    display_grid_of_all(axs['B'], sample.photo, sample.masks)
-    corners_number_dist(axs['C'], polygons)
-    area_dist(axs['D'], polygons)
-    average_angle_dist(axs["E"], sample.masks)
+    samples = []
+    directory = os.fsencode("./data")
+    for filename in os.listdir(directory):
+        filename = os.fsdecode(filename)
+        print(f"processing {filename}")
+        try:
+            sample = Sample(filename, mask_generator)
+            samples.append(sample)
+        except Exception as e:
+            print(f"error processing {filename}: {e}")
+    print(f"proccessing done")
 
-    axs['H'].axis('off')
-    axs["H"].text(0, 0, f"polygons: {len(polygons)}")
-
-    # plt.show()
-    
-    template = mpld3.fig_to_html(fig, d3_url=None, mpld3_url=None, no_extras=False, template_type='general', figid=None, use_http=False)
-    with tempfile.NamedTemporaryFile('w', delete=False, suffix='.html') as f:
-        url = 'file://' + f.name
-        f.write(template)
-    webbrowser.open(url)
+    display.display_multiple(samples)
+    print("done")
 
 if __name__ == "__main__":
     if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS'):
